2021-week1-3589: Accepted
